jd now uses new bandwidth limiter and speedmeter
git-svn-id: svn://svn.jdownloader.org/jdownloader/trunk@10393 ebf7c1c2-ba36-0410-9fe8-c592906822b4

+++ src/jd/plugins/LinkStatus.java
@@ -331,7 +331,7 @@ public class LinkStatus implements Serializable {
             removeStatus(DOWNLOADINTERFACE_IN_PROGRESS);
         }
         if (hasStatus(LinkStatus.DOWNLOADINTERFACE_IN_PROGRESS)) {
-            final int speed = downloadLink.getDownloadSpeed();
+            final long speed = downloadLink.getDownloadSpeed();
             String chunkString = "";
             if (dli != null && dli.getChunkNum() > 1) chunkString = "(" + dli.getChunksDownloading() + "/" + dli.getChunkNum() + ")";
